Added new functions DBSelect.whereGroup(groups) and DBSelect.whereGroupClose(groups).
These function will allow the creation of arbitrary parenthetical
expressions in the WHERE clause.  Each has a parameter specifying how
many groups to open or close.  It's smart enough to not close groups
that aren't open and to automatically close open groups even if you
forget to.

As an example, this:

dbWrapper.getSelect()
          .from('user')
          .whereGroupClose() // Erroneous whereGroupClose to be handled
          .whereGroup(3)     // Multiple group start
          .where('enabled=1')
          .where( 'id=?', 10 )
          .whereGroupClose()
          .where( 'first_name=?', 'Dr.')
          .whereGroup()
          .where( 'last_name LIKE ?', '%Benton%' )
          .orWhere( 'nickname=?', 'dude' )
          .whereGroupClose(2);
          // Automatic closing of leftover groups

Generates this:

"SELECT user.* FROM user WHERE (((enabled=1 AND id=10) AND
first_name='Dr.' AND (last_name LIKE '%Benton%' OR nickname='dude')))"